+++ lib/nrf_ble_dfu.dart
@@ -1,7 +1,7 @@
 library nrf_ble_dfu;
 
+export 'src/ble.dart';
 export 'src/state/state.dart';
 export 'src/enum/enum.dart';
 export 'src/extension/extension.dart';
 export 'src/nrf_ble_dfu.dart';
-