+++ lib/src/state/state.dart
@@ -5,3 +5,4 @@ export 'dfu_setup_state.dart';
 export 'dfu_preset.dart';
 export 'dfu_history_entry.dart';
 export 'log_entry.dart';
+export 'notifier.dart';