+++ test/dfu_crc32_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:nrf_ble_dfu/nrf_ble_dfu.dart';
 
 void main() {